Move NewInstanceCellEditor to Application Plugin
The NewInstanceCellEditor is and shall only be used in the application
plugin. To reduce dependencies it is moved there. Furthermore the
PaletteFilter is moved along. It is only used by the cell editor and it
reduces UI dependencies of the model plugin.
